Proper layering of Bottom Castle
Added a transparent Interaction layer that handles all mouse clicks every thing else is just pictures
