+++ mtd/src/main/java/dk/dtu/mtd/view/TowerShop.java
@@ -31,7 +31,7 @@ public class TowerShop extends HBox {
                 event.consume();
             }
         });
-        final ImageView superTower = new ImageView(new Image("dk/dtu/mtd/assets/dartMonkey.png", 100, 0, true, true));
+        final ImageView superTower = new ImageView(new Image("dk/dtu/mtd/assets/AOEtower.png", 100, 0, true, true));
         this.getChildren().add(superTower);
         superTower.addEventHandler(MouseEvent.DRAG_DETECTED, new EventHandler<MouseEvent>() {
             @Override
